add simple benchmarking notebook
This creates a simple notebook (`benchmarks/enemies/benchmark.ipynb`) for evaluating the system prompt in the Enemy agent. It includes a number of enemies as designed by the Encounter Designer agent (see `benchmarks/enemies/README.md` for more) and it implements a simple script that applies a number of questions to an Enemy agent seeded with each of them. There is no scoring to speak of, but at least it gives you a quick and easy way to evaluate the impact of a change you have made.