extend benchmarks to include longer conversations
This commit extends the more interesting examples so that it implements the full conversation instead of just providing history. For example, there is an exercise where an enemy is asked to take damage and then subsequently asked what its current HP is. We implement the test case by running the full conversation through the model, instead of supplying the chat history and only evaluating the response to the last question. This helps us get at more nuanced behaviors such as whether the agent is doing a good job keeping track of its own status.

With this change, using OpenAI's API, the suite costs about $0.22 to run and takes about 1min.