wrap benchmarks in pytest script
(still not working yet). The idea here is that one would run the suite of benchmarks using pytest. This would allow us to add asserts for some things (e.g. we definitely want it to roll dice in this situation) and continue to use our current (eyeball it) approach for evaluating the more nuanced prompt results.