e2e example with elastic embeddings
This commit includes 3 notebooks which together are sufficient to demonstrate end to end how this approach might work for an AI dungeon master.

It includes the kernel of a few ingestion strategies to bring in PDFs and embed them using OpenAI's text embedding model. One strategy simply embeds an entire chapter and the other strategy first splits it up into paragraphs and embeds each. In playing around with the results, I have found that the former strategy works better, probably because the per-paragraph strategy did not include overlap, so search results would draw from paragraphs that were semantically similar to the query, but actually contextually very dissimilar. So it was not a good tool to answer questions about D&D rules. That said, maybe a better chunking strategy (e.g. by sub-section, with overlap, etc) would yield even better results.

This commit also includes a third notebook that demonstrates the principle of how this can work with simple a simple RAG example. That is, it interprets the user query in order to construct a useful ElasticSarch query, then passes the results at inference time in order to generate a ruling. It works quite well but takes a long time to get results, so finding ways to preserve the accuracy while lowering latency will be the next order of business.