2 implementations for combat encounters
In our scratchpad, we attempted two approaches for navigating turns in combat. In one, we used a different set of developer prompts for a) determining the outcome of player actions vs b) narrating NPC actions and determining their outcome. In the second, we had a single prompt that was responsible both for resolving outcomes of player actions AND narrating/resolving NPC actions, as well as keeping track of turn order. Based on the informal experimentation we have been doing, the second approach seems to work better.