use poppler for ingestion pipeline
This commit changes the way we use Elastic. Instead of submitting raw CBOR or base64 encoded content to Elastic, for it to parse into text using the Tika engine, we use the poppler library instead, and submit the parsed, cleaned documents into a new index. This allows us to create embeddings of particular sections, rather than restricting ourselves to chapters of PDFs as we were before. It also has the nice side effect of cleaning up some garbling of the "fl" and "fi" character sequences, which were mis-evaluated under Tika.