load games from special source-books index (#11)
This adds an implementation of a `source-books` index containing all of the independent source books that have been added. This index contains the title
of the book, the game it applies to, and the name of the index into which the book has been ingested.

It also adds an endpoint that can be used to fetch the set of games for which there exist source books, and configures the frontend to use it.